Refactored JSONMessageEncoder to policies/abstract.py
