+++ arroyo/processing/strategies/dead_letter_queue/policies/produce.py
@@ -1,9 +1,7 @@
-import json
 import time
 from collections import deque
 from concurrent.futures import Future
-from datetime import datetime
-from typing import Any, Deque, Mapping, Optional
+from typing import Deque, Optional
 
 from arroyo.backends.abstract import Producer
 from arroyo.backends.kafka.consumer import KafkaPayload
@@ -11,25 +9,11 @@ from arroyo.processing.strategies.dead_letter_queue.policies.abstract import (
     DeadLetterQueuePolicy,
     InvalidMessage,
     InvalidMessages,
-    JSONSerializable,
 )
 from arroyo.types import Message, Topic
-from arroyo.utils.codecs import Encoder
 from arroyo.utils.metrics import get_metrics
 
 MAX_QUEUE_SIZE = 5000
-DATE_TIME_FORMAT = "%Y-%m-%d %H:%M:%S"
-
-
-class JSONMessageEncoder(Encoder[bytes, Mapping[str, JSONSerializable]]):
-    def __default(self, value: Any) -> str:
-        if isinstance(value, datetime):
-            return value.strftime(DATE_TIME_FORMAT)
-        else:
-            raise TypeError
-
-    def encode(self, value: Mapping[str, JSONSerializable]) -> bytes:
-        return json.dumps(value, default=self.__default).encode("utf-8")
 
 
 class ProduceInvalidMessagePolicy(DeadLetterQueuePolicy):
@@ -57,7 +41,7 @@ class ProduceInvalidMessagePolicy(DeadLetterQueuePolicy):
         self.__metrics.increment("dlq.produced_messages", len(e.messages))
 
     def _build_payload(self, message: InvalidMessage) -> KafkaPayload:
-        data = JSONMessageEncoder().encode(message.to_dict())
+        data = message.to_bytes()
         return KafkaPayload(key=None, value=data, headers=[])
 
     def _produce(self, payload: KafkaPayload) -> None:
